Not defer close of Channel when in flush
Motivation:

Before we defered the closing of the Channel when we were flushing. This is problematic as this means that if the user add a ChannelFutureListener that closed the Channel we was not excecute this until we are done with flushing. This can lead to more data send as expected.

Modifications:

- Not defer closing when in flush

Result:

Correctly respect order of events and closing the Channel ASAP